+++ packages/provider_kit/lib/src/state/index.dart
@@ -1,3 +1,4 @@
 export 'notifiers/state_notifier.dart';
 export 'type_defs/state_callbacks.dart';
+export 'state_field.dart';
 export 'widgets/index.dart';